DT-906: moved tests
