+++ tests/_data/Zed/Persistence/SpyEntityUsageRuleTest/testFacadeCanNotCreateSpyEntity.php
