+++ tests/_data/Zed/Persistence/SpyEntityUsageRuleTest/testRepositoryCanCreateSpyEntity.php
